% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bcrp_metadata.R
\name{get_bcrp_metadata}
\alias{get_bcrp_metadata}
\title{Access metadata for all available datasets provided by the Peruvian Central Bank}
\usage{
get_bcrp_metadata()
}
\value{
A data frame with one row per code available for request.
}
\description{
Access metadata for all available datasets provided by the Peruvian Central Bank
}
\examples{
# No arguments required, simply call the function.
get_bcrp_metadata()

}
\author{
Julio Collazos
}
